# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salli Islamic Prayer App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterBody {
  /** @minLength 2 */
  displayName: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UpdateProfileBody {
  /** @minLength 2 */
  displayName: string;
  /**
   * @minLength 3
   * @maxLength 30
   */
  username?: string;
  /** @nullable */
  city?: string | null;
}

export interface AuthUser {
  id: number;
  displayName: string;
  /** @nullable */
  username?: string | null;
  email: string;
  /** @nullable */
  city?: string | null;
  createdAt: string;
}

export interface PrayerDay {
  date: string;
  fajr: boolean;
  dhuhr: boolean;
  asr: boolean;
  maghrib: boolean;
  isha: boolean;
  completedCount: number;
}

export interface WeeklyPrayerDay {
  date: string;
  dayName: string;
  fajr: boolean;
  dhuhr: boolean;
  asr: boolean;
  maghrib: boolean;
  isha: boolean;
  completedCount: number;
  isToday: boolean;
}

export interface WeeklyPrayers {
  days: WeeklyPrayerDay[];
  weeklyTotal: number;
  weeklyPercentage: number;
}

export interface PrayerStats {
  currentStreak: number;
  totalPrayers: number;
  weeklyTotal: number;
  weeklyPercentage: number;
}

export type SocialUserFriendshipStatus =
  (typeof SocialUserFriendshipStatus)[keyof typeof SocialUserFriendshipStatus];

export const SocialUserFriendshipStatus = {
  none: "none",
  pending_sent: "pending_sent",
  pending_received: "pending_received",
  friends: "friends",
} as const;

export interface SocialUser {
  id: number;
  displayName: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  city?: string | null;
  friendshipStatus: SocialUserFriendshipStatus;
}

export interface FriendRequest {
  id: number;
  user: SocialUser;
  createdAt: string;
}

export interface DirectMessage {
  id: number;
  senderId: number;
  receiverId: number;
  content: string;
  /** @nullable */
  readAt?: string | null;
  createdAt: string;
}

export interface MessageThread {
  partner: SocialUser;
  lastMessage: DirectMessage;
  unreadCount: number;
}

export interface SendDirectMessageBody {
  /** @minLength 1 */
  content: string;
}

export interface NotificationsCount {
  pendingRequests: number;
  unreadMessages: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface LeaderboardEntry {
  userId: number;
  displayName: string;
  /** @nullable */
  city?: string | null;
  weeklyTotal: number;
  weeklyPercentage: number;
  currentStreak: number;
  isCurrentUser: boolean;
  rank: number;
}

export type SearchUsersParams = {
  q: string;
};